requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.2.0
django-environ==0.4.4
django-filter==1.1.0
django-oauth-toolkit==1.2.0
django-starterkit==1.0.3
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.6
oauthlib==2.1.0
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
